Test `DatasetSample` in its own test suite
